requirements.txt
requests
python-dotenv
pytest-cov
pyright
